Refactor SetupWizard, move it to the EvaluationsRootView
